change from putItem to updateItem
put is not the best option to explain DynamoDB Item update. In the earlier example, the attributes that are not updated, like createdAt will be lost, as put writes the entire new document (and the current example does not include the createdAt attribute), resulting in the attributes not supplied in the put being lost. Also updated the validation condition from AND (&&) to OR (||)